backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
dotenv
google-genai
requests
